+++ lib/serverpod_vps.dart
@@ -1,2 +1,4 @@
+export 'src/deployment_stack_config.dart';
 export 'src/docker_network_name.dart';
+export 'src/https_port_acme_warning.dart';
 export 'src/template_generator.dart';